requirements.txt

# Web Framework
Flask==3.0.0

# Configuration
PyYAML==6.0.1

# Computer Vision & Motion Detection
# NOTE: For Jetson Orin Nano Super, use the pre-built OpenCV with CUDA support
# Installation: See JETSON_SETUP.md for details
# opencv-python==4.8.1.78  # Comment out on Jetson - use system OpenCV with CUDA
numpy==1.24.3

# ONVIF Camera Discovery
onvif-zeep==0.2.12
zeep==4.2.1

# Network Monitoring & Telegram Alerts
psutil==5.9.6
requests==2.31.0

# Utilities
# ffmpeg - Install via apt on Jetson: sudo apt install ffmpeg (includes NVENC/NVDEC support)
